[driver] RW layer and image commit
Re-implement the driver interface to mount the volume with a RW layer and
commit/push the layer when the volume is removed. For the time being we can use
a hack where we create a container from the image and bind-mount the graph
driver's merged directory into the target container. Later on the docker engine
may provide the necessary primitives to union the layers and give us a
commitable rw layer.