[driver] Optionally push volume when unmounting
The main motivation for this project is to be able to mount an image as a
context in a pipeline, modify the context, and ship the new version to the next
step in the pipeline. To that end this change builds out functionality that
lets you specify a registry in the target tag of the volume which will receive
the new image when the volume is unmounted.